Created buttons for manipulating server
Added 4 button:
- Edit - redirect to page, on which configure and modified server configuration
- Delete - remove server
- Start - start server
- Stop - Stop server

Other changes:
- Added django_filter for version and server listViewAPI
- Refactoring components:
	- Created general link, which looks like default button
	- Added special function, which check if JWT token isn't expried

+++ apps/server/utils/java/server.py
@@ -82,7 +82,9 @@ class MinecraftServer:
         template(self._PATH, self._VERSION)
 
         server = Server(
-            version=Version.objects.get(version=self._VERSION), name=self._NAME, user=user
+            version=Version.objects.get(version=self._VERSION),
+            name=self._NAME,
+            user=user,
         )
         server.save()
 
@@ -138,4 +140,6 @@ class MinecraftServer:
             self.stop(id)
 
         server.delete()
-        shutil.rmtree(self._PATH)
+
+        if os.path.isdir(self._PATH):
+            shutil.rmtree(self._PATH)
